requirements.txt
matplotlib>=3.8
numpy>=1.26
pillow>=10.0
PyYAML>=6.0
scikit-learn>=1.4
seaborn>=0.13
streamlit>=1.35
torch>=2.2
torchvision>=0.17
